# specs/1-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API
  description: API for the RAG Chatbot integrated with Docusaurus Book
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /ingest:
    post:
      summary: Ingest book content from /docs directory
      description: Processes all markdown files in /docs and indexes them for RAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                force_reindex:
                  type: boolean
                  description: Whether to reindex all content even if unchanged
                  default: false
              example:
                force_reindex: false
      responses:
        '200':
          description: Ingestion completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success, error]
                  processed_files:
                    type: integer
                    description: Number of files processed
                  indexed_chunks:
                    type: integer
                    description: Number of chunks indexed
                  message:
                    type: string
                    description: Status message
                example:
                  status: success
                  processed_files: 15
                  indexed_chunks: 150
                  message: "Successfully processed 15 files and indexed 150 chunks"
        '500':
          description: Ingestion failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [error]
                  message:
                    type: string
                    description: Error message
                example:
                  status: error
                  message: "Failed to process files: [error details]"

  /chat:
    post:
      summary: Chat with full-book RAG
      description: Ask questions about the entire book content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: User's question or message
                session_id:
                  type: string
                  description: Optional session ID to maintain conversation context
              example:
                message: "What are the key concepts in chapter 3?"
                session_id: "abc123"
      responses:
        '200':
          description: Chat response with citations
          content:
            application/json:
              schema:
                type: object
                required:
                  - response
                  - session_id
                properties:
                  response:
                    type: string
                    description: The chatbot's response
                  session_id:
                    type: string
                    description: Session ID for conversation continuity
                  citations:
                    type: array
                    items:
                      type: object
                      properties:
                        source_file:
                          type: string
                          description: Path to the source file
                        title:
                          type: string
                          description: Title of the source document
                        content_snippet:
                          type: string
                          description: Snippet of the cited content
                    description: List of citations used in the response
                example:
                  response: "Chapter 3 covers the fundamental concepts of RAG systems..."
                  session_id: "abc123"
                  citations:
                    - source_file: "/docs/chapter3.md"
                      title: "Understanding RAG Systems"
                      content_snippet: "The fundamental concept of Retrieval-Augmented Generation..."
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: Error details
        '500':
          description: Chat processing failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: Error details

  /chat/selection:
    post:
      summary: Chat with selected-text RAG mode
      description: Ask questions based only on provided text selection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - selected_text
                - question
              properties:
                selected_text:
                  type: string
                  description: The text selected by the user
                question:
                  type: string
                  description: The question about the selected text
                session_id:
                  type: string
                  description: Optional session ID to maintain conversation context
              example:
                selected_text: "The fundamental concept of Retrieval-Augmented Generation combines document retrieval with language model generation to provide accurate, source-based responses."
                question: "Can you explain this concept in simpler terms?"
                session_id: "abc123"
      responses:
        '200':
          description: Chat response with citations
          content:
            application/json:
              schema:
                type: object
                required:
                  - response
                  - session_id
                properties:
                  response:
                    type: string
                    description: The chatbot's response
                  session_id:
                    type: string
                    description: Session ID for conversation continuity
                  citations:
                    type: array
                    items:
                      type: object
                      properties:
                        source_file:
                          type: string
                          description: Path to the source file
                        title:
                          type: string
                          description: Title of the source document
                        content_snippet:
                          type: string
                          description: Snippet of the cited content
                    description: List of citations used in the response
                example:
                  response: "In simple terms, RAG systems retrieve relevant documents and then use a language model to generate answers based on those documents..."
                  session_id: "abc123"
                  citations:
                    - source_file: "user_provided_text"
                      title: "Selected Text"
                      content_snippet: "The fundamental concept of Retrieval-Augmented Generation combines..."
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: Error details
        '500':
          description: Chat processing failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: Error details

  /health:
    get:
      summary: Health check
      description: Check the health status of the service and its dependencies
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - timestamp
                properties:
                  status:
                    type: string
                    enum: [healthy]
                  timestamp:
                    type: string
                    format: date-time
                    description: Current timestamp
                  services:
                    type: object
                    properties:
                      database:
                        type: string
                        description: Database connection status
                      vector_store:
                        type: string
                        description: Vector store connection status
                      openai_api:
                        type: string
                        description: OpenAI API connectivity status
                example:
                  status: healthy
                  timestamp: "2025-12-16T10:00:00Z"
                  services:
                    database: "connected"
                    vector_store: "connected"
                    openai_api: "reachable"
        '503':
          description: Service is unhealthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - timestamp
                properties:
                  status:
                    type: string
                    enum: [unhealthy]
                  timestamp:
                    type: string
                    format: date-time
                    description: Current timestamp
                  services:
                    type: object
                    properties:
                      database:
                        type: string
                        description: Database connection status
                      vector_store:
                        type: string
                        description: Vector store connection status
                      openai_api:
                        type: string
                        description: OpenAI API connectivity status
                  error:
                    type: string
                    description: Error message
                example:
                  status: unhealthy
                  timestamp: "2025-12-16T10:00:00Z"
                  services:
                    database: "disconnected"
                    vector_store: "connected"
                    openai_api: "reachable"
                  error: "Database connection failed"